+++ src/main/java/cn/tsxygfy/blog/mapper/InfoMapper.java
@@ -1,16 +1,17 @@
 package cn.tsxygfy.blog.mapper;
 
 import cn.tsxygfy.blog.model.po.Info;
+
 import java.util.List;
 
 public interface InfoMapper {
-    int deleteByPrimaryKey(Integer id);
+    Long deleteByPrimaryKey(Integer id);
 
-    int insert(Info record);
+    Long insert(Info record);
 
     Info selectByPrimaryKey(Integer id);
 
     List<Info> selectAll();
 
-    int updateByPrimaryKey(Info record);
+    Long updateByPrimaryKey(Info record);
 }
