+++ src/main/java/cn/tsxygfy/blog/mapper/LinksMapper.java
@@ -1,16 +1,17 @@
 package cn.tsxygfy.blog.mapper;
 
 import cn.tsxygfy.blog.model.po.Links;
+
 import java.util.List;
 
 public interface LinksMapper {
-    int deleteByPrimaryKey(Long id);
+    Long deleteByPrimaryKey(Long id);
 
-    int insert(Links record);
+    Long insert(Links record);
 
     Links selectByPrimaryKey(Long id);
 
     List<Links> selectAll();
 
-    int updateByPrimaryKey(Links record);
+    Long updateByPrimaryKey(Links record);
 }
