+++ src/main/java/cn/tsxygfy/blog/mapper/SettingMapper.java
@@ -1,16 +1,17 @@
 package cn.tsxygfy.blog.mapper;
 
 import cn.tsxygfy.blog.model.po.Setting;
+
 import java.util.List;
 
 public interface SettingMapper {
-    int deleteByPrimaryKey(Integer id);
+    Long deleteByPrimaryKey(Integer id);
 
-    int insert(Setting record);
+    Long insert(Setting record);
 
     Setting selectByPrimaryKey(Integer id);
 
     List<Setting> selectAll();
 
-    int updateByPrimaryKey(Setting record);
+    Long updateByPrimaryKey(Setting record);
 }
