+++ src/main/java/cn/tsxygfy/blog/mapper/SystemLogMapper.java
@@ -1,16 +1,17 @@
 package cn.tsxygfy.blog.mapper;
 
 import cn.tsxygfy.blog.model.po.SystemLog;
+
 import java.util.List;
 
 public interface SystemLogMapper {
-    int deleteByPrimaryKey(Long id);
+    Long deleteByPrimaryKey(Long id);
 
-    int insert(SystemLog record);
+    Long insert(SystemLog record);
 
     SystemLog selectByPrimaryKey(Long id);
 
     List<SystemLog> selectAll();
 
-    int updateByPrimaryKey(SystemLog record);
+    Long updateByPrimaryKey(SystemLog record);
 }
